migrations/models/4_20250404140022_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `movies` MODIFY COLUMN `external_id` INT;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `movies` MODIFY COLUMN `external_id` INT NOT NULL;"""
