pyproject.toml
[project]
name = "fastapi-assignment-hard"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = ["fastapi[standard] (>=0.115.11,<0.116.0)", "black (>=25.1.0,<26.0.0)", "ruff (>=0.9.10,<0.10.0)", "mypy (==1.13.0)", "pytest-asyncio (>=0.25.3,<0.26.0)", "coverage (>=7.6.12,<8.0.0)", "httpx (>=0.28.1,<0.29.0)", "passlib[bcrypt] (>=1.7.4,<2.0.0)", "pyjwt (>=2.10.1,<3.0.0)", "types-passlib (>=1.7.7.20250318,<2.0.0.0)", "pydantic-settings (>=2.8.1,<3.0.0)", "aerich (>=0.8.2,<0.9.0)", "tomlkit (>=0.13.2,<0.14.0)", "cryptography (>=44.0.2,<45.0.0)", "tortoise-orm[asyncmy] (==0.23.0)", "pillow (>=11.1.0,<12.0.0)"]

[[project.authors]]
name = "meoyoug"
email = "eodyd426@gmail.com"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py312"

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.12
strict = true
ignore_missing_imports = true

[tool.coverage.run]
omit = ["*/test_*.py"]

[tool.coverage.report]
exclude_also = ["def __repr__", "def __str__", "if __name__ == .__main__.:"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.aerich]
tortoise_orm = "src.configs.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
